api/requirements.txt
Flask==2.3.2
requests==2.31.0
Flask-CORS==3.0.10
flask-caching==1.10.1
redis==4.5.5
Flask-Limiter==2.6.0
python-dotenv==1.0.0